backend/pyproject.toml
[project]
name = "research-assistant-backend"
version = "0.1.0"
description = "Research Assistant Agent — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-multipart",
    "pymupdf",
    "anthropic",
    "openai",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "qdrant-client",
    "aiofiles",
    "google-genai",
    "groq",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
